+++ core/rust-lib/src/tracking/mod.rs
@@ -11,6 +11,7 @@
 #![allow(dead_code)]
 
 pub mod db;
+pub mod export;
 pub mod os;
 
 use crate::db::DbHandle;